+++ src/app/features/dashboard/all.js
@@ -1,6 +1,7 @@
 define([
   './dashboardCtrl',
   './dashboardNavCtrl',
+  './cloneDashboardCtrl',
   './playlistCtrl',
   './rowCtrl',
   './sharePanelCtrl',